chore: add image cache module
